refactor: move some methods to utils
